// ui-bikes-app/src/app/services/models/bike-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BikeResponse {
  bikeName?: string;
  borrowed?: boolean;
  description?: string;
  favorite?: boolean;
  id?: number;
  ownerName?: string;
  picture?: Array<string>;
  rating?: number;
  repairing?: boolean;
  shareable?: boolean;
  techIdentifier?: string;
}